move date to common
